made prediction time based rather than frame based
