+++ source/panel_finder/panel_finder.py
@@ -1,4 +1,3 @@
-from source.panel_finder.find_center import find_target_center, find_dict_center
 from source.panel_finder.led_finder.led_finder import LEDFinder
 from source.common.module import Module
 from pathlib import Path
@@ -13,6 +12,18 @@ def combined_panel(rect_a, rect_b):
     new_rect = {"x_center": center[0], "y_center": center[1], "width": width, "height": height, "angle": angle}
 
     return new_rect
+
+
+def find_dict_center(bounds):
+    """
+        finds the center of 2 rectangles
+
+        :param bounds: 2 rectangles in the dictionary form
+        :return: the center in the form [x,y]
+        """
+    targetcX = int((bounds[0]["x_center"] + bounds[1]["x_center"]) / 2.0)
+    targetcY = int((bounds[0]["y_center"] + bounds[1]["y_center"]) / 2.0)
+    return targetcX, targetcY
 
 
 class PanelFinder(Module):
@@ -30,12 +41,16 @@ class PanelFinder(Module):
         super().__init__(self.working_dir, state=state)
         self.led_finder = LEDFinder()
         self.panel = None
-            
 
     def predict_leds(self, led_a, led_b, frame_dims):
         return self.classifier.process((led_a, led_b), frame_dims)
 
     def process(self, frame):
+        """
+        find a robot panel from and image
+        :param frame: an image that may contain a robot
+        :return: a dict describing a rotated rectangle
+        """
         frame_dims = frame.shape[:2]
         leds = self.led_finder.process(frame)
         leds = sorted(leds, key=lambda x: x['angle'])
